py/requirements.txt
supabase
python-dotenv
PyPDF2
langchain
langchain-groq
flask
flask-cors